refactor(params): make `LibEVMVersion` a constant
